renamed json_validation to validate
